exceptions: add accessors for pc and error
in order to share most of the fault code in core/mmu.cc,
we add accessors to get the program counter and an error
code whose interpretation is arch-dependent
(error_code for x64, ESR for AArch64).

This also fixes an issue in jvm_balloon, where the
error code was checked incorrectly.

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>