sched: add explicit realtime policy state
trying to guess the exact policy as set with sched_setscheduler
from the current realtime state is not possible, and makes the
code unoptimal.
This change moves the setting of the non-RT priority inside
the set_realtime, which now deals with setting the right one
according to the passed POSIX policy.

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>