pci: enable bars issuing pci command
for architectures other than x64 we need to reprogram the
BAR registers as we have no magical good values already
programmed by the "firmware" (qemu...).
It is also possible that there will be the need to reprogram
them on x64 in the future.
Trying to do that without explicitly enabling the I/O space
and memory space using a PCI command leads, in the case of
QEMU, to the reprogramming to be silently ignored.

There is therefore some minor restructuring of the pci
device / function code to account for a slightly different
procedure for the BARs, where we now:

enable access to I/O and mem, then
for each bar
    we first read the bar size, then
    call some arch-dependent code to fix up the final BAR reg,
    so that archs have a chance to reprogram it, then
    we proceed to assign the final address to the bar.

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>