aarch64: mmu: extend device memory range
at some point we need to get rid of those hardcoded addresses,
but that requires some changes to propagate the property through
the MMU template code, which I still didn't manage to work
reliably.

Suppress the debug message which was emitted for each page
with memory attributes for device memory, since when adding the
full PCI space for mach-virt that was starting to pollute the
output too much (and it seems to be working fine anyway).

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>
Reviewed-by: Nadav Har'El <nyh@cloudius-systems.com>