zfs-tool: extract zfs-userspace from zfs-fuse to zfs-tool
to clearly distinguish the zfs-fuse package from the zfs-tool,
move the code as to leave the original package as unchanged as possible
(with just the small #include modification needed to compile),
and bolt it in with the build system.

This requires some changes to opensolaris_atomic.cc in order
to be compilable independently on the host_arch, and to provide the
function symbols expected by the zfs libraries included in zfs-fuse.

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>