aarch64: setup secondary CPUs
get the information about available cpus and mpids from the devicetree.

Add a per-cpu initial stack, using the same mechanism as for x64.

Adjust boot.S for the secondary cpus, copying over the ttbr0 and ttbr1
settings from the boot cpu.

In smp.cc set up the arch-cpu structure with the initial stack,
the cpu index (necessary because of GICv2) and the MPIDR, then launch
the cpus in smp.cc using the PSCI CPU_ON functionality.

Signed-off-by: Jani Kokkonen <jani.kokkonen@huawei.com>

[claudio: some rework, bugfixing]

Signed-off-by: Claudio Fontana <claudio.fontana@huawei.com>